feat: Add image property to CreateBoard schema
This commit adds the image property to the CreateBoard object schema, which is used to validate the input for creating a new board. The image property is a string that is required and has no minimum length. The commit also adds custom error messages for the image property, similar to the title property.